+++ examples/cpp-gen/tasklist.obx.hpp
@@ -9,15 +9,32 @@
 #include "objectbox.h"
 #include "objectbox.hpp"
 
+
 struct Task_;
 
 struct Task {
-    using _OBX_MetaInfo = Task_;
-    
-    uint64_t id;
+    obx_id id;
     std::string text;
     uint64_t date_created;
     uint64_t date_finished;
+
+    struct _OBX_MetaInfo {
+        static constexpr obx_schema_id entityId() { return 1; }
+    
+        static void setObjectId(Task& object, obx_id newId) { object.id = newId; }
+    
+        /// Write given object to the FlatBufferBuilder
+        static void toFlatBuffer(flatbuffers::FlatBufferBuilder& fbb, const Task& object);
+    
+        /// Read an object from a valid FlatBuffer
+        static Task fromFlatBuffer(const void* data, size_t size);
+    
+        /// Read an object from a valid FlatBuffer
+        static std::unique_ptr<Task> newFromFlatBuffer(const void* data, size_t size);
+    
+        /// Read an object from a valid FlatBuffer
+        static void fromFlatBuffer(const void* data, size_t size, Task& outObject);
+    };
 };
 
 struct Task_ {
@@ -25,21 +42,5 @@ struct Task_ {
     static const obx::Property<Task, OBXPropertyType_String> text;
     static const obx::Property<Task, OBXPropertyType_Date> date_created;
     static const obx::Property<Task, OBXPropertyType_Date> date_finished;
-
-    static constexpr obx_schema_id entityId() { return 1; }
-
-    static void setObjectId(Task& object, obx_id newId) { object.id = newId; }
-
-    /// Write given object to the FlatBufferBuilder
-    static void toFlatBuffer(flatbuffers::FlatBufferBuilder& fbb, const Task& object);
-
-    /// Read an object from a valid FlatBuffer
-    static Task fromFlatBuffer(const void* data, size_t size);
-
-    /// Read an object from a valid FlatBuffer
-    static std::unique_ptr<Task> newFromFlatBuffer(const void* data, size_t size);
-
-    /// Read an object from a valid FlatBuffer
-    static void fromFlatBuffer(const void* data, size_t size, Task& outObject);
 };