+++ doxygen/Changelog.md
@@ -3,6 +3,13 @@
 ObjectBox C and C++ API Changelog
 =================================
 
+0.13.0 (2021-03-16)
+-------------------
+* split Sync symbols out of objectbox.h/pp into objectbox-sync.h/pp 
+* add Sync server-time getter, listener and local-to-server diff info - to access server time info on clients 
+* add Sync heartbeat interval configuration and an option to send one immediately
+* semi-internal: update Dart/Flutter SDK to v2.0
+
 0.12.0 (2021-02-05)
 -------------------
 * add Linux ARMv8 (aarch64) native binary library